+++ capstone/capstone-grp1/backend/src/main/java/com/group1/capstone/repository/CargoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.group1.capstone.model.Cargo;
 
 @Repository
-public interface CargoRepository extends JpaRepository<Cargo, Long> {
+public interface CargoRepository extends JpaRepository<Cargo, Integer> {
     
 }
