+++ capstone/capstone-grp1/backend/src/main/java/com/group1/capstone/repository/PortRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.group1.capstone.model.Port;
 
 @Repository
-public interface PortRepository extends JpaRepository<Port, Long> {
+public interface PortRepository extends JpaRepository<Port, Integer> {
     
 }
