+++ capstone/capstone-grp1/backend/src/main/java/com/group1/capstone/repository/TrackingEventRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.group1.capstone.model.TrackingEvent;
 
 @Repository
-public interface TrackingEventRepository extends JpaRepository<TrackingEvent, Long> {
+public interface TrackingEventRepository extends JpaRepository<TrackingEvent, Integer> {
     
 }
